--- a/Project.toml
+++ b/Project.toml
@@ -20,7 +20,7 @@ StatsFuns = "4c63d2b9-4356-54db-8cca-17b64c39e42c"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [compat]
-CliqueTrees = "0.4"
+CliqueTrees = "0.5"
 DataStructures = "0.18"
 Distributions = "0.25"
 ForwardDiff = "0.10"