@samuelsonric 2025-02-23
Here is a [larger example](https://github.com/ropensci/phylotaR/blob/0495c37ed0a21878193ee1080a108d9ce8cbf74d/other/plants.tre).

```julia
julia> graph.graph
{8467, 8466} undirected simple Int16 graph
```

**current implementation**
```julia
julia> @benchmark PGBP.triangulate_minfill!(copy($graph))
BenchmarkTools.Trial: 5 samples with 1 evaluation per sample.
 Range (min … max):  1.212 s …  1.233 s  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.223 s             ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.223 s ± 7.622 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █                         █  █  █                      █  
  █▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁█▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  1.21 s        Histogram: frequency by time        1.23 s <

 Memory estimate: 8.42 MiB, allocs estimate: 126148.
 ```
 
 **new implementation**
 
 ```julia
 julia> @benchmark PGBP.fast_triangulate_minfill!(copy($graph))
 BenchmarkTools.Trial: 844 samples with 1 evaluation per sample.
 Range (min … max):  5.428 ms …  11.034 ms  ┊ GC (min … max): 0.00% … 47.70%
 Time  (median):     5.649 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.921 ms ± 690.360 μs  ┊ GC (mean ± σ):  4.75% ±  8.57%

  ▁▄▆██▇▆▄▁                          ▂▁▁                       
  █████████▇▇▆▁▁▅▁▄▁▄▁▄▁▁▁▁▁▁▁▅▆████▇████▅▇▆██▅█▅▄▁▅▁▁▄▆▁▄▁▁▄ █
  5.43 ms      Histogram: log(frequency) by time      8.12 ms <

 Memory estimate: 5.83 MiB, allocs estimate: 75921.
 ```