@samuelsonric 2025-03-03
@cecileane @bstkj Any problems, here? If not, I might submit another PR optimizing the function `cliquetree`.