@cecileane 2025-03-04
The only problem is my lack time to review your PR! (teaching, committees, other deadlines).
@bstkj is also working on a quite major update to the package, so he will need to check that your PR has no conflict with his branch.

My only concern is that your package brings a somewhat large number of dependencies.
The clique tree calculation is not the primary bottleneck for what this package is doing. The bottleneck that would be most beneficial to optimize is belief propagation, which is called over and over for parameter estimation. Building the clique tree only happens once. But your PR speeds up this first step enormously indeed!